fix index type in two examples

+++ examples/cvodes/F2003_serial/cvsAdvDiff_FSA_non_f2003.f90
@@ -78,9 +78,9 @@ module ode_problem
   real(c_double),  parameter :: ATOL  = 1e-5
   integer(c_int),  parameter :: NOUT  = 10
   integer(c_int),  parameter :: NP    = 2
-  integer(c_int),  parameter :: NS = 2
-  integer(c_long), parameter :: MX    = 10
-  integer(c_long), parameter :: NEQ   = MX
+  integer(c_int),  parameter :: NS    = 2
+  integer(kind=myindextype), parameter :: MX  = 10
+  integer(kind=myindextype), parameter :: NEQ = MX
 
   ! problem constants
   real(c_double) :: ZERO  = 0.d0
